requirements.txt
requests==2.25.1
XlsxWriter==1.4.0
requests_cache==0.6.3
beautifulsoup4==4.9.3
Pillow==8.2.0